% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_R.R
\name{fn_a}
\alias{fn_a}
\alias{fn_b}
\title{Test file}
\usage{
fn_a(a)

fn_b(b)
}
\arguments{
\item{a}{test param number 1}

\item{b}{test param number 2}
}
\description{
Test file
}
\examples{
testpkg:::fn_a(4)
testpkg:::fn_b("world")
hist(rnorm(1000))
library(rbokeh)
}
\keyword{programming}

